requirements.txt
instaloader==4.10.3
instagrapi==2.0.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
google-generativeai==0.3.2
yt-dlp==2023.12.30
moviepy==1.0.3
python-dateutil==2.8.2
flask==2.3.3
gunicorn==21.2.0
numpy<2
